backend/pyproject.toml
[project]
name = "skating-backend"
version = "0.1.0"
description = "FastAPI backend for figure skating biomechanics analysis"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30",
    "bcrypt>=4.0,<4.1",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "passlib[bcrypt]>=1.7",
    "PyJWT>=2.9",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.22",
    "redis>=5.0.0",
    "sse-starlette>=3.3.4",
    "sqlalchemy[asyncio]>=2.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.42.0",
    "alembic>=1.18.4",
    "arq>=0.27.0",
    "boto3>=1.42.83",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
